added comments declaration and variables is more structured, according features usage

+++ ble5_simple_peripheral_cc26x2r1lp_app/Application/project_zero.c
@@ -162,7 +162,7 @@ PIN_Config ledPinTable[] = {
     PIN_TERMINATE
 };
 
-extern int stream_on;
+extern uint8_t stream_on;
 bool connection_occured = FALSE;
 /*********************************************************************
  * LOCAL FUNCTIONS
